pyproject.toml
[project]
name = "basic-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
    "gradio>=4.0.0",
    "langchain-anthropic>=0.3.13",
    "langchain-openai>=0.3.17",
    "langchain[google-genai]>=0.1.0",
    "langgraph>=0.0.20",
    "langsmith>=0.3.42",
    "openai>=1.6.0",
    "pyaudio>=0.2.13",
    "python-dotenv>=1.0.0",
    "speechrecognition>=3.10.0",
    "typing-extensions>=4.5.0",
    "whisper>=1.1.10",
]
